Make virtual scroll table rows be uniform height
This fixes a problem that prevents all results from being loaded if the
first table row is shorter than most of the rest of the table rows.

One example of such a table is one that includes image thumbnails that
exceed the height of the text columns, but the thumbnails don't exist
in all rows. Another example is a table that contains some cells that
have text that wraps to multiple lines (either through a CSS override
that sets white-space:normal on the cell or HTML content that renders
as multiple lines).